requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.37.2
megfile==3.1.2
tiktoken
verovio
opencv-python
accelerate
numpy==1.26.4
shutils
pillow
python-dotenv